Fix time_until_ready overflow on extreme banked energy
time_until_ready computed ACTION_COST - energy and deficit + speed - 1
in i32, which panics in debug (and wraps in release) once set_energy
banks a value near i32::MIN — the same reachable state next_turn already
defends with i64 math. Widen to i64; the return type becomes i64 since
the true wait can exceed i32::MAX.

+++ izanagi_kit/tests/properties.rs
@@ -4224,7 +4224,11 @@ fn prop_scheduler_time_until_ready_formula() {
             } else {
                 (deficit + speed - 1) / speed
             };
-            assert_eq!(s.time_until_ready(id), Some(expected), "formula mismatch");
+            assert_eq!(
+                s.time_until_ready(id),
+                Some(expected as i64),
+                "formula mismatch"
+            );
             assert_eq!(
                 s.time_until_ready(id) == Some(0),
                 energy >= ACTION_COST,
